Add GraphQL infrastructure
Adds an API for GraphQL requests, and a build script to autogenerate TS types from a GraphQL schema + GraphQL queries.

The most commonly used GraphQL client seems to be Apollo, but I'm not convinced it's not overkill for this usecase. In order to keep the dependencies lightweight but also keep our options open in the future, I've instead written a simple shim compatible with the codegen intended for Apollo but not actually using Apollo as a dependency.

Since the GraphQL API doesn't get called yet, this PR is fully backwards compatible.